Move createStripeMethod to custom hook
